WIP on VideoSwitch Ross communicator

+++ Plugins/TAS.Client.VideoSwitch/Model/Interfaces/IVideoSwitchCommunicator.cs
@@ -6,7 +6,7 @@ namespace TAS.Server.VideoSwitch.Model.Interfaces
     {
         void Preload(int sourceId);     
         void SetTransitionStyle(VideoSwitcherTransitionStyle mixEffect);
-        void SetMixSpeed(double rate);
+        void SetMixSpeed(byte rate);
         void Take();
     }
 }
